refactor(render): type synthetic series fields honestly, extract scatter spec builder
Per review: syntheticSeriesField is now a narrow SyntheticSeriesField
interface covering exactly what the spec generators read (name, getLabel,
valueSet, referenceId, maxString), so the bar and line plugins construct
it directly with no cast. The six is*() predicates the old literals faked
were read nowhere and are gone. The generators' series variable is typed
Field | SyntheticSeriesField.

Scatter's spec construction moved verbatim out of renderToDOM into
generate-scatter_chart-spec.ts so it can be unit-tested without vega's
ESM-only View; the render path is unchanged. While moving it, the
color/size/shape titles switched from the non-standard axis property to
the documented legend property; the compiled output is identical (pinned
by the new spec test and verified in Storybook).

The duplicated tooltip entry-label expression in each generator is hoisted
into one entryLabel closure so both tooltip branches share the tested
implementation.
