+++ packages/malloy-render/src/plugins/bar-chart/bar-chart-plugin.tsx
@@ -38,12 +38,13 @@ import {
   barChartSettingsSchema,
 } from './bar-chart-settings';
 import {barChartSettingsToTag} from './settings-to-tag';
+import type {SyntheticSeriesField} from '@/plugins/synthetic-series-field';
 
 export interface BarChartPluginInstance extends CoreVizPluginInstance<BarChartPluginMetadata> {
   getTopNSeries?: (maxSeries: number) => (string | number | boolean)[];
   field: NestField;
   chartDisplay: ChartDisplayConfig;
-  syntheticSeriesField?: Field;
+  syntheticSeriesField?: SyntheticSeriesField;
   hasMultipleSeriesFields?: boolean;
 }
 
@@ -196,20 +197,12 @@ export const BarChartPluginFactory: RenderPluginFactory<BarChartPluginInstance> 
           }
 
           if (hasMultipleSeriesFields) {
-            // Create synthetic field
             pluginInstance.syntheticSeriesField = {
               name: seriesFields.map(f => f.name).join(' - '),
               getLabel: () => seriesFields.map(f => f.getLabel()).join(' - '),
               valueSet: concatenatedValues,
               referenceId: '__synthetic_concatenated_series__',
-              // Minimal Field interface implementation
-              isTime: () => false,
-              isDate: () => false,
-              isBasic: () => true,
-              isNumber: () => false,
-              isString: () => true,
-              isBoolean: () => false,
-            } as unknown as Field;
+            };
           }
         },
